[fix] 添加账号设置 related to issue #1

+++ components/logout/index.tsx
@@ -6,13 +6,16 @@ import cookie from '../cookie'
 type Props = RouteComponentProps<{}>
 class Main extends React.Component<Props> {
   public componentWillMount () {
-    userLogout()
-    Config.token = ''
-    Config.user = undefined
-    localStorage.clear()
-    sessionStorage.clear()
-    cookie.removeAll()
-    Config.history('/login')
+    userLogout().then(() => {
+      Config.token = ''
+      Config.user = undefined
+      if (window.localStorage) {
+        localStorage.clear()
+        sessionStorage.clear()
+      }
+      cookie.removeAll()
+      Config.history('/login')
+    })
   }
   public render () {
     return (
